Remove JS Toolbox, add adsupport page, update LinkedIn link and setup basic preloader in index page
With this update, JS Toolbox is removed from being a part of the main website. It has more scope, and is planned to be launched as a separate project (repository) later. A new adsupport page is included, so as to try including Google AdSense to support ads on the page, for those open to support website that way. The LinkedIn Handle Public Link is also updated to match with the one set same as social profile. The preloader is currently used as a text loader in just home page, and is to be improved later.