Redesigned Page Preloader, Add 404.html, Update about page content and skill tags content
With this commit, the page preloader is redesigned, and is applied to the index and about page for now. It is planned to be applied to each and every page soon, in the coming commits. A 404 Error Page addition has been attempted using 404.html file. About page content has also been enhanced a little bit.